% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category.R
\docType{class}
\name{Category}
\alias{Category}
\title{Category}
\format{An \code{R6Class} generator object}
\usage{
Category
}
\description{
Category Class
}
\section{Fields}{

\describe{
\item{\code{id}}{integer  [optional]}

\item{\code{name}}{character  [optional]}
}}

\keyword{datasets}
