% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiResponse.r, R/Category.r, R/Order.r,
%   R/Tag.r, R/User.r, R/pet.R
\docType{data}
\name{Element}
\alias{Element}
\alias{Element}
\alias{Element}
\alias{Element}
\alias{Element}
\alias{Element}
\title{Element Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Element

Element

Element

Element

Element

Element
}
\description{
Element Class

Element Class

Element Class

Element Class

Element Class

Element Class
}
\keyword{datasets}
