% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_api_response.R
\docType{class}
\name{ModelApiResponse}
\alias{ModelApiResponse}
\title{ModelApiResponse}
\format{An \code{R6Class} generator object}
\usage{
ModelApiResponse
}
\description{
ModelApiResponse Class
}
\section{Fields}{

\describe{
\item{\code{code}}{integer  [optional]}

\item{\code{type}}{character  [optional]}

\item{\code{message}}{character  [optional]}
}}

\keyword{datasets}
