% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\docType{class}
\name{Order}
\alias{Order}
\title{Order}
\format{An \code{R6Class} generator object}
\usage{
Order
}
\description{
Order Class
}
\section{Fields}{

\describe{
\item{\code{id}}{integer  [optional]}

\item{\code{petId}}{integer  [optional]}

\item{\code{quantity}}{integer  [optional]}

\item{\code{shipDate}}{character  [optional]}

\item{\code{status}}{character  [optional]}

\item{\code{complete}}{character  [optional]}
}}

\keyword{datasets}
