% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet.R
\docType{class}
\name{Pet}
\alias{Pet}
\title{Pet}
\format{An \code{R6Class} generator object}
\usage{
Pet
}
\description{
Pet Class
}
\section{Fields}{

\describe{
\item{\code{id}}{integer  [optional]}

\item{\code{category}}{\link[petstore:Category]{  Category  }  [optional]}

\item{\code{name}}{character}

\item{\code{photoUrls}}{list(character)}

\item{\code{tags}}{\link[petstore:Tag]{   list(Tag)   }  [optional]}

\item{\code{status}}{character  [optional]}
}}

\keyword{datasets}
