% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petstore.R
\docType{data}
\name{PetStoreClient}
\alias{PetStoreClient}
\title{Pet Store Client Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PetStoreClient
}
\description{
Pet Store Client Class
}
\keyword{datasets}
