% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiResponse.r, R/Category.r, R/Order.r,
%   R/Tag.r, R/User.r, R/pet.R
\docType{data}
\name{Response}
\alias{Response}
\alias{Response}
\alias{Response}
\alias{Response}
\alias{Response}
\alias{Response}
\title{Response Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Response

Response

Response

Response

Response

Response
}
\description{
Response Class

Response Class

Response Class

Response Class

Response Class

Response Class
}
\keyword{datasets}
