% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\docType{class}
\name{Tag}
\alias{Tag}
\title{Tag}
\format{An \code{R6Class} generator object}
\usage{
Tag
}
\description{
Tag Class
}
\section{Fields}{

\describe{
\item{\code{id}}{integer  [optional]}

\item{\code{name}}{character  [optional]}
}}

\keyword{datasets}
