% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\docType{class}
\name{User}
\alias{User}
\title{User}
\format{An \code{R6Class} generator object}
\usage{
User
}
\description{
User Class
}
\section{Fields}{

\describe{
\item{\code{id}}{integer  [optional]}

\item{\code{username}}{character  [optional]}

\item{\code{firstName}}{character  [optional]}

\item{\code{lastName}}{character  [optional]}

\item{\code{email}}{character  [optional]}

\item{\code{password}}{character  [optional]}

\item{\code{phone}}{character  [optional]}

\item{\code{userStatus}}{integer  [optional]}
}}

\keyword{datasets}
